+++ pkg/runner/step_context.go
@@ -77,7 +77,7 @@ func (sc *StepContext) Executor(ctx context.Context) common.Executor {
 		remoteAction.URL = rc.Config.GitHubInstance
 
 		github := rc.getGithubContext()
-		if remoteAction.IsCheckout() && github.isLocalCheckout(step) {
+		if remoteAction.IsCheckout() && isLocalCheckout(github, step) {
 			return func(ctx context.Context) error {
 				common.Logger(ctx).Debugf("Skipping local actions/checkout because workdir was already copied")
 				return nil
